Fix IFS (input field separator) for each function
When in the code there is a change in IFS, then the functions can lead
to errors when splitting the strings with spaces into arrays.
For this reason, each time we enter into a function we need to save IFS
to an old var and the restored once we exit the function.

Signed-off-by: EasyNet <devel@easynet.dev>